chore: clean up XOR stark
